ai_service_quick/requirements.txt
pandas
pandas-ta
numpy==1.26.4
pytest
python-dotenv
scipy
fastapi
uvicorn[standard]
scikit-learn>=1.4.0
kaggle
kagglehub
lightgbm